Various cleanups and fixes

+++ examples/server/server.h
@@ -29,8 +29,7 @@ THREAD_RETURN WOLFSSL_THREAD server_test(void* args);
 /* Echo bytes using buffer of blockSize until [echoData] bytes are complete. */
 /* If [bechmarkThroughput] set the statistcs will be output at the end */
 int ServerEchoData(WOLFSSL* ssl, int clientfd, int echoData, int blockSize,
-                   int benchmarkThroughput);
+                   size_t benchmarkThroughput);
 
 
 #endif /* WOLFSSL_SERVER_H */
-
